+++ database/migrations/2023_04_02_114129_create_budgets_table.php
@@ -16,9 +16,9 @@ return new class extends Migration
             //'bank','type','number','budget','last_date'
             $table->string('bank');
             $table->string('type');
-            $table->unsignedInteger('user_id');
+            $table->unsignedBigInteger('user_id');
             $table->char('numbers',4);
-            $table->unsignedInteger('budget');
+            $table->unsignedBigInteger('budget');
             $table->char('last_date',4);
             $table->timestamps();
         });
